(svn r195) Auto-correction code for nvar in variational 2s and nrand in random 2s.

+++ pseudo.h
@@ -48,7 +48,7 @@ public:
 	PseudoSprite&SetDec(uint,uint){return*this;}
 	PseudoSprite&SetBE(uint,uint){return*this;}
 
-	PseudoSprite&SetByteAt(uint,uint);
+	PseudoSprite&SetByteAt(uint,uchar);
 
 	PseudoSprite&PadAfter(uint,uint);
